Refactor Gif class
