+++ src/core/mykeyboard.cpp
@@ -604,7 +604,7 @@ String keyboard(String mytext, int maxSize, String msg) {
   }
 
   //Resets screen when finished writing
-  tft.fillRect(0,0,tftWidth,tftHeight,bruceConfig.bgColor);
+  tft.fillScreen(bruceConfig.bgColor);
   resetTftDisplay();
 
   return mytext;
